+++ modules/app-desktop/src/main/kotlin/org/archivekeep/app/desktop/ui/dialogs/storages/MarkAsLocalDialog.kt
@@ -5,17 +5,12 @@ import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.launch
 import org.archivekeep.app.core.domain.storages.KnownStorage
 import org.archivekeep.app.core.domain.storages.Storage
 import org.archivekeep.app.core.persistence.platform.demo.asKnownStorage
@@ -24,8 +19,13 @@ import org.archivekeep.app.core.persistence.registry.RegistryDataStore
 import org.archivekeep.app.core.utils.identifiers.StorageURI
 import org.archivekeep.app.desktop.domain.wiring.LocalRegistry
 import org.archivekeep.app.desktop.ui.components.dialogs.SimpleActionDialogControlButtons
+import org.archivekeep.app.desktop.ui.components.dialogs.operations.LaunchableExecutionErrorIfPresent
 import org.archivekeep.app.desktop.ui.designsystem.dialog.DialogPreviewColumn
 import org.archivekeep.app.desktop.ui.utils.appendBoldSpan
+import org.archivekeep.app.desktop.utils.Launchable
+import org.archivekeep.app.desktop.utils.asTrivialAction
+import org.archivekeep.app.desktop.utils.mockLaunchable
+import org.archivekeep.app.desktop.utils.simpleLaunchable
 import org.archivekeep.utils.loading.Loadable
 
 class MarkAsLocalDialog(
@@ -37,28 +37,24 @@ class MarkAsLocalDialog(
         val storage: KnownStorage,
         val onClose: () -> Unit,
     ) {
-        var runningJob by mutableStateOf<Job?>(null)
-
-        fun launch() {
-            runningJob =
-                coroutineScope.launch {
-                    registry.updateStorage(
-                        storage.storageURI,
-                    ) {
-                        it.copy(
-                            isLocal = true,
-                        )
-                    }
-                    onClose()
+        val action =
+            simpleLaunchable(coroutineScope) {
+                registry.updateStorage(
+                    storage.storageURI,
+                ) {
+                    it.copy(isLocal = true)
                 }
-        }
+                onClose()
+            }
     }
 
     class State(
         val storage: KnownStorage,
-        val onLaunch: () -> Unit,
+        val launchable: Launchable<Unit>,
     ) : IState {
         override val title = buildAnnotatedString { append("Mark storage as local") }
+
+        val action = launchable.asTrivialAction()
     }
 
     @Composable
@@ -81,7 +77,7 @@ class MarkAsLocalDialog(
             Loadable.Loaded(
                 State(
                     vm.storage,
-                    vm::launch,
+                    vm.action,
                 ),
             )
         }
@@ -97,6 +93,7 @@ class MarkAsLocalDialog(
                     append(" as local.")
                 },
         )
+        LaunchableExecutionErrorIfPresent(state.launchable)
     }
 
     @Composable
@@ -106,7 +103,7 @@ class MarkAsLocalDialog(
     ) {
         SimpleActionDialogControlButtons(
             "Mark as local",
-            onLaunch = state.onLaunch,
+            actionState = state.action.value,
             onClose = onClose,
         )
     }
@@ -121,7 +118,7 @@ private fun preview1() {
         dialog.renderDialogCardForPreview(
             MarkAsLocalDialog.State(
                 hddA.asKnownStorage(),
-                onLaunch = {},
+                mockLaunchable(false, null),
             ),
         )
     }