+++ modules/app-desktop/src/main/kotlin/org/archivekeep/app/desktop/utils/SingleLaunchGuard.kt
@@ -1,39 +1,37 @@
 package org.archivekeep.app.desktop.utils
 
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
+import org.archivekeep.app.core.utils.generics.ExecutionOutcome
 
-class LaunchableAction(
+class SingleLaunchGuard(
     val scope: CoroutineScope,
 ) {
     var runningJob by mutableStateOf<Job?>(null)
+        private set
 
-    val isRunning
-        get() = runningJob != null
+    var executionOutcome = mutableStateOf<ExecutionOutcome?>(null)
 
     fun launch(fn: suspend () -> Unit) {
         runningJob =
             scope.launch {
                 try {
                     fn()
+
+                    executionOutcome.value = ExecutionOutcome.Success()
+                } catch (e: Throwable) {
+                    executionOutcome.value = ExecutionOutcome.Failed(e)
                 } finally {
                     runningJob = null
                 }
             }
     }
-}
 
-@Composable
-fun rememberLaunchableAction(scope: CoroutineScope = rememberCoroutineScope()) =
-    remember {
-        LaunchableAction(
-            scope = scope,
-        )
+    fun reset() {
+        executionOutcome.value = null
     }
+}